Renamed files of periodType to ticketRepresentation
